requirements.txt
django==4.2.2
djangorestframework==3.14.0
django-filter==23.2
drf_yasg==1.21.6
coreapi==2.3.3


Pillow==9.5.0
